pyproject.toml
[project]
name = "my-custom-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.1"
dependencies = [
    "boto3>=1.42.16",
    "botocore",
    "fastapi>=0.127.0",
    "httpx>=0.28.1",
    "logging>=0.4.9.6",
    "pydantic>=2.12.5",
    "strands-agents>=1.20.0",
    "uvicorn[standard]>=0.40.0",
]

[tool.setuptools]
py-modules = [
  "agent",
  "deploy_agent",
  "invoke_agent",
  "main"
]
